Update test names and small refactoring
