+++ packages/wallet-backend-update/src/app.ts
@@ -13,7 +13,7 @@ import type { AuthService } from './auth/service'
 import type { Env } from './config/env'
 import type { UserService } from './user/service'
 import type { UserController } from './user/controller'
-import { Container } from './container'
+import { Container } from './shared/container'
 import { Model } from 'objection'
 import { withSession } from './middleware/withSession'
 import { errorHandler } from './middleware/errorHandler'
