+++ packages/wallet-backend-update/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "lib": ["ESNext"],
-    "rootDir": "./src",
     "sourceMap": true,
     "baseUrl": ".",
     "paths": {
